requirements.txt
rich[jupyter]
privex-helpers
privex-loghelper
pyyaml

